contractshield-python/pyproject.toml
[project]
name = "contractshield"
version = "0.1.0"
description = "Contract-First API Security for Python"
authors = [{name = "AuditLabs", email = "contact@contractshield.dev"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["api", "security", "validation", "openapi", "fastapi", "contract"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]

dependencies = [
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.104.0", "starlette>=0.27.0"]
flask = ["flask>=3.0.0"]
cel = ["cel-python>=0.1.5"]
all = [
    "fastapi>=0.104.0",
    "starlette>=0.27.0",
    "flask>=3.0.0",
    "cel-python>=0.1.5",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://contractshield.dev"
Documentation = "https://docs.contractshield.dev"
Repository = "https://github.com/AuditLabs/contractshield-python"
Issues = "https://github.com/AuditLabs/contractshield-python/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contractshield"]

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in argument defaults
]

[tool.ruff.isort]
known-first-party = ["contractshield"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/contractshield --cov-report=term-missing"

[tool.coverage.run]
source = ["src/contractshield"]
branch = true
